# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Canvas API",
    "description": "RESTful API for LLM Canvas operations",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/canvas": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Canvas",
        "description": "Get a full canvas by ID.\nArgs:\n    canvas_id: Canvas UUID to retrieve\nReturns:\n    CanvasData on success\nRaises:\n    HTTPException: 404 if canvas not found",
        "operationId": "get_canvas_api_v1_canvas_get",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canvas UUID",
              "title": "Canvas Id"
            },
            "description": "Canvas UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCanvasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create Canvas",
        "description": "Create a new canvas.\nArgs:\n    request: Canvas creation request with optional title and description\nReturns:\n    CreateCanvasResponse with the canvas ID and success message",
        "operationId": "create_canvas_api_v1_canvas_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCanvasRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCanvasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/canvas/{canvas_id}": {
      "delete": {
        "tags": [
          "v1"
        ],
        "summary": "Delete Canvas",
        "description": "Delete a canvas by ID.\nArgs:\n    canvas_id: Canvas UUID to delete\nReturns:\n    DeleteCanvasResponse with success message\nRaises:\n    HTTPException: 404 if canvas not found",
        "operationId": "delete_canvas_api_v1_canvas__canvas_id__delete",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canvas UUID to delete",
              "title": "Canvas Id"
            },
            "description": "Canvas UUID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCanvasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/canvas/{canvas_id}/messages": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Commit Message",
        "description": "Commit a new message to a canvas.\nArgs:\n    canvas_id: Canvas UUID to add message to\n    request: Canvas commit message event data\nReturns:\n    CreateMessageResponse with the message ID and success message\nRaises:\n    HTTPException: 404 if canvas not found",
        "operationId": "commit_message_api_v1_canvas__canvas_id__messages_post",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canvas UUID",
              "title": "Canvas Id"
            },
            "description": "Canvas UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/canvas/{canvas_id}/messages/{message_id}": {
      "put": {
        "tags": [
          "v1"
        ],
        "summary": "Update Message",
        "description": "Update an existing message in a canvas.\nArgs:\n    canvas_id: Canvas UUID containing the message\n    message_id: Message ID to update\n    request: Canvas update message event data\nReturns:\n    CreateMessageResponse with the message ID and success message\nRaises:\n    HTTPException: 404 if canvas or message not found",
        "operationId": "update_message_api_v1_canvas__canvas_id__messages__message_id__put",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canvas UUID",
              "title": "Canvas Id"
            },
            "description": "Canvas UUID"
          },
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Message ID to update",
              "title": "Message Id"
            },
            "description": "Message ID to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/canvas/{canvas_id}/sse": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Canvas Message Sse",
        "description": "Server-Sent Events endpoint for canvas message updates.\n\nSends events when messages are added, updated, or deleted in a specific canvas.\nEvents include:\n- message_committed: When a new message is added to the canvas\n- message_updated: When an existing message is updated\n- message_deleted: When a message is deleted\n\nArgs:\n    canvas_id: Canvas UUID to stream events for\nReturns:\n    StreamingResponse with SSE events\nRaises:\n    HTTPException: 404 if canvas not found",
        "operationId": "canvas_message_sse_api_v1_canvas__canvas_id__sse_get",
        "parameters": [
          {
            "name": "canvas_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canvas UUID",
              "title": "Canvas Id"
            },
            "description": "Canvas UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CanvasCommitMessageEvent": {
        "properties": {
          "event_type": {
            "type": "string",
            "const": "commit_message",
            "title": "Event Type"
          },
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/MessageNode"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "canvas_id",
          "timestamp",
          "data"
        ],
        "title": "CanvasCommitMessageEvent",
        "description": "Event data for canvas message commits."
      },
      "CanvasData": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MessageNode"
            },
            "type": "object",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "title",
          "last_updated",
          "description",
          "canvas_id",
          "created_at",
          "nodes"
        ],
        "title": "CanvasData",
        "description": "Complete canvas data structure."
      },
...
      "CanvasUpdateMessageEvent": {
        "properties": {
          "event_type": {
            "type": "string",
            "const": "update_message",
            "title": "Event Type"
          },
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/MessageNode"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "canvas_id",
          "timestamp",
          "data"
        ],
        "title": "CanvasUpdateMessageEvent",
        "description": "Event data for canvas message updates."
      },
...
      "CommitMessageRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CanvasCommitMessageEvent"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CommitMessageRequest"
      },
      "CreateCanvasRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CreateCanvasRequest",
        "description": "Request type for POST /api/v1/canvas"
      },
      "CreateCanvasResponse": {
        "properties": {
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "canvas_id",
          "message"
        ],
        "title": "CreateCanvasResponse",
        "description": "Response type for POST /api/v1/canvas"
      },
      "CreateMessageResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "canvas_id",
          "message"
        ],
        "title": "CreateMessageResponse",
        "description": "Response type for POST /api/v1/canvas/{canvas_id}/messages"
      },
      "DeleteCanvasResponse": {
        "properties": {
          "canvas_id": {
            "type": "string",
            "title": "Canvas Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "canvas_id",
          "message"
        ],
        "title": "DeleteCanvasResponse",
        "description": "Response type for DELETE /api/v1/canvas/{canvas_id}"
      },
      "GetCanvasResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CanvasData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "GetCanvasResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextBlockParam"
                    },
                    {
                      "$ref": "#/components/schemas/ToolUseBlockParam"
                    },
                    {
                      "$ref": "#/components/schemas/ToolResultBlockParam"
                    },
                    {
                      "$ref": "#/components/schemas/ImageBlockParam"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Content"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "content",
          "role"
        ],
        "title": "Message",
        "description": "Message structure for canvas conversations."
      },
      "MessageNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "child_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Child Ids"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message",
          "parent_id",
          "child_ids",
          "meta"
        ],
        "title": "MessageNode",
        "description": "Node in the canvas conversation graph."
      },
...
      "TextBlockParam": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "type": {
            "type": "string",
            "const": "text",
            "title": "Type"
          },
          "cache_control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CacheControlEphemeralParam"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CitationCharLocationParam"
                    },
                    {
                      "$ref": "#/components/schemas/CitationPageLocationParam"
                    },
                    {
                      "$ref": "#/components/schemas/CitationContentBlockLocationParam"
                    },
                    {
                      "$ref": "#/components/schemas/CitationWebSearchResultLocationParam"
                    },
                    {
                      "$ref": "#/components/schemas/CitationSearchResultLocationParam"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "text",
          "type"
        ],
        "title": "TextBlockParam"
      },
      "ToolResultBlockParam": {
        "properties": {
          "tool_use_id": {
            "type": "string",
            "title": "Tool Use Id"
          },
          "type": {
            "type": "string",
            "const": "tool_result",
            "title": "Type"
          },
          "cache_control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CacheControlEphemeralParam"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextBlockParam"
                    },
                    {
                      "$ref": "#/components/schemas/ImageBlockParam"
                    },
                    {
                      "$ref": "#/components/schemas/SearchResultBlockParam"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Content"
          },
          "is_error": {
            "type": "boolean",
            "title": "Is Error"
          }
        },
        "type": "object",
        "required": [
          "tool_use_id",
          "type"
        ],
        "title": "ToolResultBlockParam"
      },
      "ToolUseBlockParam": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "input": {
            "title": "Input"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "const": "tool_use",
            "title": "Type"
          },
          "cache_control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CacheControlEphemeralParam"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "input",
          "name",
          "type"
        ],
        "title": "ToolUseBlockParam"
      },
...
      "UpdateMessageRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CanvasUpdateMessageEvent"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "UpdateMessageRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}